requirements.txt
anthropic>=0.8.0
openai>=1.0.0
google-generativeai>=0.3.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
requests>=2.31.0
newsapi-python>=0.2.7 